Changed int to float for thickness and diameter
Changed int to float for thickness and diameter (millsizetype, pipe,
connector)
Issue: PRISM-51 Настройки-Труба-Параметры типоразмера- Диаметр трубы и
Толщина стенки должны быть вещественными

+++ src/Domain/Entity/Setup/PipeMillSizeType.cs
@@ -16,8 +16,8 @@ namespace Prizm.Domain.Entity.Setup
         }
         public virtual string Type { get; set; }
         public virtual int Length { get; set; }
-        public virtual int Diameter { get; set; }
-        public virtual int Thickness { get; set; }
+        public virtual float Diameter { get; set; }
+        public virtual float Thickness { get; set; }
         public virtual bool IsNative { get; set; }
         public virtual Project Project { get; set; }
